requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.2
json2html==1.3.0
Pillow==9.5.0
pyOpenSSL==23.1.1
pystray==0.19.4
python_dateutil==2.8.2
Requests==2.30.0
scikit_learn==1.2.2
whois==0.9.27
win_inet_pton==1.1.0
windows_toasts==0.4.1
googlesearch-python==1.2.3